Refactor to try shape up report a bit on having a notion of errors by column and errors by row as two distinct collections.
In case of columns only columns that have errors are listed.

However with rows, all validated data rows are included but some
field items in row may have a null error element indicating the data in
that field in the row is valid.

Not totally sure this is the best design, but as rows is a list of
rows, we can map back to original data by row number more easily
if we keep all rows.

We also persist the original input data and the subset of columns
that we actually validated (as we actually ignore the deprecated and
unknown columns).

Separated out the report into the report class and validation report
class. The latter is a persistent class outside of the validation
module. It is used only in the flask app to persist validation
results to a database.

This is a breaking change and will require the drop and
recreation of databases.