Hooked up update header flow and download modified csv.
Changes persisted so user can revalidate.

Lots of renames and refactoring to make things a little clearer,
not totally successful as code is still quite confusing about
what changes are make and keeping lists of additional headers,
missing headers etc.

TODO - The download of csv relates just to the data we
extract from original upload. However we still need
to add back in the deprecated headers that we ignore for
purposes of validation.